create delete produk

+++ pages/api/produk/delete.js
@@ -1,18 +1,28 @@
 import { prisma } from "../../../libs/prisma.libs";
 
-export default function handler(req, res) {
-    prisma.product.delete({
+export default async function handler(req, res) {
+  const { id } = req.query;
+  if (id && req.method === "DELETE") {
+    try {
+      const product = await prisma.product.delete({
         where: {
-            id: req.body.id,
-        }
-    }).then((produk) => {
-        res.status(200).json({
-            message: "Produk deleted!",
-            data: produk,
-        });
-    }).catch((error) => {
-        res.status(500).json({
-            message: error.message || "Error occured! Please contact the admin for more information.",
-        });
+          id: parseInt(id),
+        },
+      });
+      res.status(201).json({
+        message: "Delete product success",
+        data: product,
+      });
+    } catch (error) {
+      res.status(400).json({
+        message: "Delete product failed",
+        data: [],
+      });
+    }
+  } else {
+    res.status(400).json({
+      message: "Method not allowed",
+      data: [],
     });
+  }
 }
